refactor: split live batch and schema upgrade helpers
Move live-batch helper types and parsing utilities out of the daemon batch executor, split the new focused tests out of oversized unit files, and collapse repeated schema upgrade-tail planning.

This keeps the #845 follow-up under the repository file-size budgets without changing the daemon behavior or schema semantics.

Ref #845
