+++ polylogue/storage/sqlite/schema_bootstrap.py
@@ -31,18 +31,13 @@ _RAW_SOURCE_MTIME_INDEX_SQL = (
     "ON raw_conversations(source_path, file_mtime) "
     "WHERE file_mtime IS NOT NULL"
 )
-
 _MESSAGE_TYPE_INDEX_SQL = (
     "CREATE INDEX IF NOT EXISTS idx_messages_conversation_message_type ON messages(conversation_id, message_type)"
 )
 _ATTACHMENT_REFS_MESSAGE_INDEX_SQL = (
     "CREATE INDEX IF NOT EXISTS idx_attachment_refs_message ON attachment_refs(message_id) WHERE message_id IS NOT NULL"
 )
-_PROVIDER_EVENTS_SOURCE_MESSAGE_INDEX_SQL = (
-    "CREATE INDEX IF NOT EXISTS idx_provider_events_source_message "
-    "ON provider_events(source_message_id) "
-    "WHERE source_message_id IS NOT NULL"
-)
+_PROVIDER_EVENTS_SOURCE_MESSAGE_INDEX_SQL = "CREATE INDEX IF NOT EXISTS idx_provider_events_source_message ON provider_events(source_message_id) WHERE source_message_id IS NOT NULL"
 
 
 @dataclass(frozen=True)
@@ -829,22 +824,7 @@ def build_v2_to_current_upgrade_plan(snapshot: SchemaSnapshot) -> SchemaExtensio
     """Build the declared v2 upgrade path to the current archive version."""
     v2_to_v3 = build_v2_to_v3_upgrade_plan(snapshot)
     v3_to_v4 = build_v3_to_v4_upgrade_plan(snapshot)
-    extra_statements: tuple[str, ...] = ()
-    if SCHEMA_VERSION >= 5:
-        from polylogue.storage.sqlite.schema_ddl_archive import BLOB_LEASE_DDL, TAGS_M2M_DDL
-        from polylogue.storage.sqlite.schema_ddl_cursor import SOURCE_FILE_CURSOR_DDL
-
-        extra_statements = _split_ddl_into_statements(SOURCE_FILE_CURSOR_DDL, TAGS_M2M_DDL, BLOB_LEASE_DDL)
-    if SCHEMA_VERSION >= 6:
-        from polylogue.storage.sqlite.schema_ddl_identity import IDENTITY_DDL
-
-        extra_statements = (*extra_statements, *_split_ddl_into_statements(IDENTITY_DDL))
-    if SCHEMA_VERSION >= 7:
-        extra_statements = (*extra_statements, *build_v6_to_v7_upgrade_plan(snapshot).statements)
-    if SCHEMA_VERSION >= 8:
-        extra_statements = (*extra_statements, *build_v7_to_v8_upgrade_plan(snapshot).statements)
-    if SCHEMA_VERSION >= 9:
-        extra_statements = (*extra_statements, *build_v8_to_v9_upgrade_plan(snapshot).statements)
+    extra_statements = _upgrade_tail_statements(snapshot, from_version=4)
     return SchemaExtensionPlan(
         statements=(*v2_to_v3.statements, *v3_to_v4.statements, *extra_statements),
         scripts=(),
@@ -887,6 +867,26 @@ def build_v8_to_v9_upgrade_plan(snapshot: SchemaSnapshot) -> SchemaExtensionPlan
     return SchemaExtensionPlan(statements=tuple(statements), scripts=())
 
 
+def _upgrade_tail_statements(snapshot: SchemaSnapshot, *, from_version: int) -> tuple[str, ...]:
+    statements: tuple[str, ...] = ()
+    if from_version < 5 <= SCHEMA_VERSION:
+        from polylogue.storage.sqlite.schema_ddl_archive import BLOB_LEASE_DDL, TAGS_M2M_DDL
+        from polylogue.storage.sqlite.schema_ddl_cursor import SOURCE_FILE_CURSOR_DDL
+
+        statements = _split_ddl_into_statements(SOURCE_FILE_CURSOR_DDL, TAGS_M2M_DDL, BLOB_LEASE_DDL)
+    if from_version < 6 <= SCHEMA_VERSION:
+        from polylogue.storage.sqlite.schema_ddl_identity import IDENTITY_DDL
+
+        statements = (*statements, *_split_ddl_into_statements(IDENTITY_DDL))
+    if from_version < 7 <= SCHEMA_VERSION:
+        statements = (*statements, *build_v6_to_v7_upgrade_plan(snapshot).statements)
+    if from_version < 8 <= SCHEMA_VERSION:
+        statements = (*statements, *build_v7_to_v8_upgrade_plan(snapshot).statements)
+    if from_version < 9 <= SCHEMA_VERSION:
+        statements = (*statements, *build_v8_to_v9_upgrade_plan(snapshot).statements)
+    return statements
+
+
 def decide_schema_bootstrap(snapshot: SchemaSnapshot) -> SchemaBootstrapDecision:
     """Choose the canonical schema bootstrap path for sync and async backends."""
     if snapshot.current_version == 0:
@@ -903,21 +903,8 @@ def decide_schema_bootstrap(snapshot: SchemaSnapshot) -> SchemaBootstrapDecision
 
     if snapshot.current_version == 3 and SCHEMA_VERSION >= 4:
         plan = build_v3_to_v4_upgrade_plan(snapshot)
-        if SCHEMA_VERSION >= 5:
-            from polylogue.storage.sqlite.schema_ddl_archive import BLOB_LEASE_DDL, TAGS_M2M_DDL
-            from polylogue.storage.sqlite.schema_ddl_cursor import SOURCE_FILE_CURSOR_DDL
-
-            extra_stmts = _split_ddl_into_statements(SOURCE_FILE_CURSOR_DDL, TAGS_M2M_DDL, BLOB_LEASE_DDL)
-            if SCHEMA_VERSION >= 6:
-                from polylogue.storage.sqlite.schema_ddl_identity import IDENTITY_DDL
-
-                extra_stmts = (*extra_stmts, *_split_ddl_into_statements(IDENTITY_DDL))
-            if SCHEMA_VERSION >= 7:
-                extra_stmts = (*extra_stmts, *build_v6_to_v7_upgrade_plan(snapshot).statements)
-            if SCHEMA_VERSION >= 8:
-                extra_stmts = (*extra_stmts, *build_v7_to_v8_upgrade_plan(snapshot).statements)
-            if SCHEMA_VERSION >= 9:
-                extra_stmts = (*extra_stmts, *build_v8_to_v9_upgrade_plan(snapshot).statements)
+        extra_stmts = _upgrade_tail_statements(snapshot, from_version=4)
+        if extra_stmts:
             plan = SchemaExtensionPlan(
                 statements=(*plan.statements, *extra_stmts),
                 scripts=(),
@@ -929,22 +916,8 @@ def decide_schema_bootstrap(snapshot: SchemaSnapshot) -> SchemaBootstrapDecision
         )
 
     if snapshot.current_version == 4 and SCHEMA_VERSION >= 5:
-        from polylogue.storage.sqlite.schema_ddl_archive import BLOB_LEASE_DDL, TAGS_M2M_DDL
-        from polylogue.storage.sqlite.schema_ddl_cursor import SOURCE_FILE_CURSOR_DDL
-
-        plan_stmts = _split_ddl_into_statements(SOURCE_FILE_CURSOR_DDL, TAGS_M2M_DDL, BLOB_LEASE_DDL)
-        if SCHEMA_VERSION >= 6:
-            from polylogue.storage.sqlite.schema_ddl_identity import IDENTITY_DDL
-
-            plan_stmts = (*plan_stmts, *_split_ddl_into_statements(IDENTITY_DDL))
-        if SCHEMA_VERSION >= 7:
-            plan_stmts = (*plan_stmts, *build_v6_to_v7_upgrade_plan(snapshot).statements)
-        if SCHEMA_VERSION >= 8:
-            plan_stmts = (*plan_stmts, *build_v7_to_v8_upgrade_plan(snapshot).statements)
-        if SCHEMA_VERSION >= 9:
-            plan_stmts = (*plan_stmts, *build_v8_to_v9_upgrade_plan(snapshot).statements)
         plan = SchemaExtensionPlan(
-            statements=plan_stmts,
+            statements=_upgrade_tail_statements(snapshot, from_version=4),
             scripts=(),
         )
         return SchemaBootstrapDecision(
@@ -954,17 +927,8 @@ def decide_schema_bootstrap(snapshot: SchemaSnapshot) -> SchemaBootstrapDecision
         )
 
     if snapshot.current_version == 5 and SCHEMA_VERSION >= 6:
-        from polylogue.storage.sqlite.schema_ddl_identity import IDENTITY_DDL
-
-        plan_statements = _split_ddl_into_statements(IDENTITY_DDL)
-        if SCHEMA_VERSION >= 7:
-            plan_statements = (*plan_statements, *build_v6_to_v7_upgrade_plan(snapshot).statements)
-        if SCHEMA_VERSION >= 8:
-            plan_statements = (*plan_statements, *build_v7_to_v8_upgrade_plan(snapshot).statements)
-        if SCHEMA_VERSION >= 9:
-            plan_statements = (*plan_statements, *build_v8_to_v9_upgrade_plan(snapshot).statements)
         plan = SchemaExtensionPlan(
-            statements=plan_statements,
+            statements=_upgrade_tail_statements(snapshot, from_version=5),
             scripts=(),
         )
         return SchemaBootstrapDecision(
@@ -974,33 +938,22 @@ def decide_schema_bootstrap(snapshot: SchemaSnapshot) -> SchemaBootstrapDecision
         )
 
     if snapshot.current_version == 6 and SCHEMA_VERSION >= 7:
-        plan = build_v6_to_v7_upgrade_plan(snapshot)
-        if SCHEMA_VERSION >= 8:
-            plan = SchemaExtensionPlan(
-                statements=(*plan.statements, *build_v7_to_v8_upgrade_plan(snapshot).statements),
-                scripts=(),
-            )
-        if SCHEMA_VERSION >= 9:
-            plan = SchemaExtensionPlan(
-                statements=(*plan.statements, *build_v8_to_v9_upgrade_plan(snapshot).statements),
-                scripts=(),
-            )
         return SchemaBootstrapDecision(
             action="upgrade_v6_to_v7",
-            extension_plan=plan,
+            extension_plan=SchemaExtensionPlan(
+                statements=_upgrade_tail_statements(snapshot, from_version=6),
+                scripts=(),
+            ),
             current_version=snapshot.current_version,
         )
 
     if snapshot.current_version == 7 and SCHEMA_VERSION >= 8:
-        plan = build_v7_to_v8_upgrade_plan(snapshot)
-        if SCHEMA_VERSION >= 9:
-            plan = SchemaExtensionPlan(
-                statements=(*plan.statements, *build_v8_to_v9_upgrade_plan(snapshot).statements),
-                scripts=(),
-            )
         return SchemaBootstrapDecision(
             action="upgrade_v7_to_v8",
-            extension_plan=plan,
+            extension_plan=SchemaExtensionPlan(
+                statements=_upgrade_tail_statements(snapshot, from_version=7),
+                scripts=(),
+            ),
             current_version=snapshot.current_version,
         )
