+++ tests/unit/sources/test_live_watcher.py
@@ -370,59 +370,6 @@ async def test_live_full_ingest_streams_large_paths_before_processing(
     assert calls == ["path:large-session.jsonl"]
 
 
-def test_full_ingest_heartbeats_small_file_groups_with_current_path(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    root = tmp_path / "sessions"
-    root.mkdir()
-    first = root / "first.jsonl"
-    second = root / "second.jsonl"
-    first.write_text('{"type":"session_meta","payload":{"id":"first"}}\n', encoding="utf-8")
-    second.write_text('{"type":"session_meta","payload":{"id":"second"}}\n', encoding="utf-8")
-    db_path = tmp_path / "archive.sqlite"
-    polylogue = SimpleNamespace(archive_root=tmp_path, backend=SimpleNamespace(db_path=db_path))
-    cursor = CursorStore(db_path)
-    processor = LiveBatchProcessor(
-        cast(Any, polylogue),
-        (WatchSource(name="codex", root=root),),
-        cursor=cursor,
-        parser_fingerprint="test-parser",
-    )
-    events: list[tuple[str, Path | None, int | None]] = []
-
-    def heartbeat(
-        phase: str,
-        *,
-        current_path: Path | None = None,
-        source_payload_read_bytes: int | None = None,
-        force: bool = False,
-    ) -> None:
-        del force
-        events.append((phase, current_path, source_payload_read_bytes))
-
-    def fake_write_from_bytes(_store: object, payload: bytes) -> tuple[str, int]:
-        suffix = f"{len(events):064x}"[-64:]
-        return suffix, len(payload)
-
-    monkeypatch.setattr(
-        "polylogue.sources.live.batch._jsonl_provider_and_conversation_artifact",
-        lambda _path, fallback_provider: (fallback_provider, True),
-    )
-    monkeypatch.setattr("polylogue.sources.live.batch.BlobStore.write_from_bytes", fake_write_from_bytes)
-    monkeypatch.setattr(
-        "polylogue.sources.live.batch._process_ingest_batch_sync",
-        lambda *args, **kwargs: SimpleNamespace(failed_raw_ids=set(), parse_failures=0, worker_count=1),
-    )
-
-    result = processor._ingest_full_paths_sync([first, second], source_name="codex", heartbeat=heartbeat)
-
-    assert result.succeeded == [first, second]
-    assert ("full_file_scan", first, 0) in events
-    assert ("full_file_scan", second, first.stat().st_size) in events
-    assert any(event == ("full_worker_wait", second, first.stat().st_size + second.stat().st_size) for event in events)
-
-
 # --- LiveWatcher: needs_work + ingest_files (batched) --------------------------
 
 
@@ -838,7 +785,7 @@ def test_full_parse_progress_groups_bounds_small_files_by_count(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     paths = [tmp_path / f"{index}.jsonl" for index in range(_SMALL_FULL_PARSE_PROGRESS_MAX_FILES + 1)]
-    monkeypatch.setattr("polylogue.sources.live.batch._path_size", lambda path: 1)
+    monkeypatch.setattr("polylogue.sources.live.batch_support._path_size", lambda path: 1)
 
     groups = list(_full_parse_progress_groups(paths))
 
@@ -851,7 +798,7 @@ def test_full_parse_progress_groups_bounds_small_files_by_bytes(
 ) -> None:
     paths = [tmp_path / f"{index}.jsonl" for index in range(5)]
     byte_size = _LARGE_FULL_PARSE_PROGRESS_BYTES - 1
-    monkeypatch.setattr("polylogue.sources.live.batch._path_size", lambda path: byte_size)
+    monkeypatch.setattr("polylogue.sources.live.batch_support._path_size", lambda path: byte_size)
 
     groups = list(_full_parse_progress_groups(paths))
